+++ src/elements/GeoBorderedToken/GeoBorderedToken.vue
@@ -1,5 +1,10 @@
 <template>
-  <div class="geo-bordered-token">
+  <div
+    class="geo-bordered-token"
+    :class="{
+      'geo-bordered-token--disabled': disabled
+    }"
+  >
     <font-awesome-icon
       v-if="icon"
       :icon="icon"
@@ -33,6 +38,16 @@ export default {
      */
     icon: {
       type: Array,
+      required: false
+    },
+
+    /**
+     * Boolean to disable the token
+     *
+     */
+
+    disabled: {
+      type: Boolean,
       required: false
     },
     /**